+++ src/app/modules/search-parameter/autocomplete-parameter-value.component.ts
@@ -20,13 +20,15 @@ import { escapeStringForRegExp, getNextPageUrl } from '../../shared/utils';
 import { catchError, expand } from 'rxjs/operators';
 import { AutocompleteParameterValue } from '../../types/autocomplete-parameter-value';
 import { HttpClient } from '@angular/common/http';
+import { LiveAnnouncer } from '@angular/cdk/a11y';
+import { FhirBackendService } from '../../shared/fhir-backend/fhir-backend.service';
 import ValueSetExpansionContains = fhir.ValueSetExpansionContains;
 import Bundle = fhir.Bundle;
-import { LiveAnnouncer } from '@angular/cdk/a11y';
 import Resource = fhir.Resource;
 import * as fhirpath from 'fhirpath';
 import * as fhirPathModelR4 from 'fhirpath/fhir-context/r4';
 import Coding = fhir.Coding;
+import { ResearchStudyService } from '../../shared/research-study/research-study.service';
 
 /**
  * data type used for this control
@@ -85,6 +87,34 @@ export class AutocompleteParameterValueComponent
     );
   }
 
+  /**
+   * Whether DbGap server is selected
+   */
+  get isDbgap(): boolean {
+    return (
+      this.fhirBackend.serviceBaseUrl ===
+      'https://dbgap-api.ncbi.nlm.nih.gov/fhir/x1'
+    );
+  }
+
+  constructor(
+    @Optional() @Self() ngControl: NgControl,
+    private elementRef: ElementRef,
+    private errorStateMatcher: ErrorStateMatcher,
+    private httpClient: HttpClient,
+    private liveAnnoncer: LiveAnnouncer,
+    private fhirBackend: FhirBackendService,
+    private researchStudy: ResearchStudyService
+  ) {
+    super();
+    if (ngControl != null) {
+      this.ngControl = ngControl;
+      // Setting the value accessor directly (instead of using
+      // the providers) to avoid running into a circular import.
+      ngControl.valueAccessor = this;
+    }
+  }
+
   static idPrefix = 'autocomplete-test-value-';
   static idIndex = 0;
   static codeTextFieldMapping = {
@@ -95,7 +125,7 @@ export class AutocompleteParameterValueComponent
   inputId =
     AutocompleteParameterValueComponent.idPrefix +
     ++AutocompleteParameterValueComponent.idIndex;
-  @Input() options: Lookup[];
+  @Input() options: Lookup[] = [];
   @Input() placeholder = '';
   @Input() resourceType: string;
   @Input() observationCodes: string[];
@@ -106,6 +136,8 @@ export class AutocompleteParameterValueComponent
   @Input() expression: string;
   @Input() usePrefetch = false;
 
+  EVIDENCEVARIABLE = 'EvidenceVariable';
+  dbgapLoincOnly = false;
   currentData: AutocompleteParameterValue = {
     codes: [],
     items: []
@@ -140,6 +172,15 @@ export class AutocompleteParameterValueComponent
   readonly disabled: boolean = false;
   readonly id: string;
   readonly required = false;
+
+  /**
+   * Returns EV id from a DbGap variable API response
+   * e.g. phv00054122.v1.p1 => phv00054122
+   * @private
+   */
+  private static getEvIdFromDbgapVariableApi(value: string): string {
+    return /^(.+)\.v\d+\.p\d+$/.test(value) ? RegExp.$1 : null;
+  }
   setDescribedByIds(): void {}
 
   /**
@@ -178,22 +219,6 @@ export class AutocompleteParameterValueComponent
     }
   }
 
-  constructor(
-    @Optional() @Self() ngControl: NgControl,
-    private elementRef: ElementRef,
-    private errorStateMatcher: ErrorStateMatcher,
-    private httpClient: HttpClient,
-    private liveAnnoncer: LiveAnnouncer
-  ) {
-    super();
-    if (ngControl != null) {
-      this.ngControl = ngControl;
-      // Setting the value accessor directly (instead of using
-      // the providers) to avoid running into a circular import.
-      ngControl.valueAccessor = this;
-    }
-  }
-
   ngOnChanges(): void {
     if (this.acInstance) {
       this.setupAutocomplete();
@@ -206,14 +231,14 @@ export class AutocompleteParameterValueComponent
 
   /**
    * Set up Autocompleter.
-   * It could be a Prefetch or a Search instance depending on this.usePrefetch.
    * Also call this.onChange() of ControlValueAccessor interface on selection event,
    * so that form control value is updated and can be read from parent form.
    */
   setupAutocomplete(): void {
-    this.acInstance = this.usePrefetch
-      ? this.setupAutocompletePrefetch()
-      : this.setupAutocompleteSearch();
+    this.acInstance =
+      this.resourceType === this.EVIDENCEVARIABLE
+        ? this.getAutocomplete_EV()
+        : this.getAutocomplete();
 
     // Fill autocomplete with data (if currentData was set in writeValue).
     this.currentData.items.forEach((item, index) => {
@@ -233,9 +258,29 @@ export class AutocompleteParameterValueComponent
   }
 
   /**
+   * Get autocompleter instance.
+   * It could be a Prefetch or a Search instance depending on this.usePrefetch.
+   */
+  getAutocomplete(): any {
+    return this.usePrefetch
+      ? this.setupAutocompletePrefetch()
+      : this.setupAutocompleteSearch();
+  }
+
+  /**
+   * Get autocompleter instance for Evidence Variable.
+   * The instance uses DbGap variable API if server is DbGap, otherwise it uses fhir queries.
+   */
+  getAutocomplete_EV(): any {
+    return this.isDbgap
+      ? this.setupAutocomplete_EV_DbgapVariableApi()
+      : this.setupAutocompleteSearch_EV();
+  }
+
+  /**
    * Set up Autocompleter prefetch options.
    */
-  setupAutocompletePrefetch(): void {
+  setupAutocompletePrefetch(): any {
     return new Def.Autocompleter.Prefetch(
       this.inputId,
       this.options.map((o) => o.display),
@@ -246,7 +291,7 @@ export class AutocompleteParameterValueComponent
   /**
    * Set up Autocompleter search options.
    */
-  setupAutocompleteSearch(): void {
+  setupAutocompleteSearch(): any {
     const acInstance = new Def.Autocompleter.Search(this.inputId, null, {
       suggestionMode: Def.Autocompleter.NO_COMPLETION_SUGGESTIONS,
       fhir: {
@@ -262,7 +307,17 @@ export class AutocompleteParameterValueComponent
               };
 
               const filterText = fieldVal;
-              if (filterText) {
+              if (
+                filterText &&
+                !(
+                  // DocumentReference.contenttype does not support querying codes by ':text'.
+                  // We will return the whole list with ':not=zzz' and filter in client.
+                  (
+                    this.resourceType === 'DocumentReference' &&
+                    this.searchParameter === 'contenttype'
+                  )
+                )
+              ) {
                 params[`${this.searchParameter}:text`] = filterText;
               } else {
                 params[`${this.searchParameter}:not`] = 'zzz';
@@ -363,6 +418,243 @@ export class AutocompleteParameterValueComponent
   }
 
   /**
+   * Set up Autocompleter search options for DbGap variable API search.
+   */
+  setupAutocomplete_EV_DbgapVariableApi(): any {
+    const acInstance = new Def.Autocompleter.Search(this.inputId, null, {
+      suggestionMode: Def.Autocompleter.NO_COMPLETION_SUGGESTIONS,
+      fhir: {
+        search: (fieldVal, count) => {
+          return {
+            then: (resolve, reject) => {
+              const url =
+                'https://clinicaltables.nlm.nih.gov/api/dbg_vars/v3/search';
+              const params = {
+                rec_type: 'dbgv',
+                terms: fieldVal,
+                maxList: count,
+                sf: `dbgv.${this.searchParameter}`,
+                df: `dbgv.${this.searchParameter}`,
+                q: this.getDbgapEvResearchStudyParam()
+              };
+              if (this.dbgapLoincOnly) {
+                params['q'] += ' has_loinc:true';
+              }
+              // Array of result items for autocompleter
+              const contains: ValueSetExpansionContains[] = [];
+              // Already selected items
+              const selectedCodes = acInstance.getSelectedItems();
+
+              this.loading = true;
+              this.subscription?.unsubscribe();
+
+              this.subscription = this.httpClient
+                .get(url, {
+                  params
+                })
+                .pipe(
+                  catchError((error) => {
+                    this.loading = false;
+                    reject(error);
+                    throw error;
+                  })
+                )
+                .subscribe((response) => {
+                  contains.push(
+                    ...this.getAutocompleteItems_EV_dbgapVariableApi(
+                      response,
+                      selectedCodes
+                    )
+                  );
+                  this.loading = false;
+                  this.liveAnnoncer.announce('Finished loading list.');
+                  resolve({
+                    resourceType: 'ValueSet',
+                    expansion: {
+                      total: response[0],
+                      contains
+                    }
+                  });
+                });
+            }
+          };
+        }
+      },
+      useResultCache: false,
+      maxSelect: '*',
+      matchListValue: true,
+      showListOnFocusIfEmpty: true
+    });
+    return acInstance;
+  }
+
+  /**
+   * Set up Autocompleter search options.
+   */
+  setupAutocompleteSearch_EV(): any {
+    const acInstance = new Def.Autocompleter.Search(this.inputId, null, {
+      suggestionMode: Def.Autocompleter.NO_COMPLETION_SUGGESTIONS,
+      fhir: {
+        search: (fieldVal, count) => {
+          return {
+            then: (resolve, reject) => {
+              const url = `$fhir/${this.EVIDENCEVARIABLE}`;
+              const params = {
+                _elements: this.searchParameter
+              };
+              params[this.searchParameter] = fieldVal;
+              // Hash of processed codes, used to exclude repeated codes
+              const processedCodes = {};
+              // Array of result items for autocompleter
+              const contains: ValueSetExpansionContains[] = [];
+              // Total amount of items
+              let total = null;
+              // Already selected items
+              const selectedCodes = acInstance.getSelectedItems();
+
+              this.loading = true;
+              this.subscription?.unsubscribe();
+
+              const obs = this.httpClient
+                .get(url, {
+                  params
+                })
+                .pipe(
+                  expand((response: Bundle) => {
+                    contains.push(
+                      ...this.getAutocompleteItems_EV(
+                        response,
+                        processedCodes,
+                        selectedCodes
+                      )
+                    );
+                    const nextPageUrl = getNextPageUrl(response);
+                    if (nextPageUrl && contains.length < count) {
+                      this.liveAnnoncer.announce('New items added to list.');
+                      // Update list before calling server for next query.
+                      resolve({
+                        resourceType: 'ValueSet',
+                        expansion: {
+                          total: Number.isInteger(total) ? total : null,
+                          contains
+                        }
+                      });
+                      const newParams = { ...params };
+                      newParams['_id:not'] = Object.keys(processedCodes).join(
+                        ','
+                      );
+                      return this.httpClient.get(url, {
+                        params: newParams
+                      });
+                    } else {
+                      if (!nextPageUrl) {
+                        total = contains.length;
+                      } else if (response.total) {
+                        total = response.total;
+                      }
+                      if (contains.length > count) {
+                        contains.length = count;
+                      }
+                      this.loading = false;
+                      this.liveAnnoncer.announce('Finished loading list.');
+                      resolve({
+                        resourceType: 'ValueSet',
+                        expansion: {
+                          total: Number.isInteger(total) ? total : null,
+                          contains
+                        }
+                      });
+                      // Emit a complete notification
+                      return EMPTY;
+                    }
+                  }),
+                  catchError((error) => {
+                    this.loading = false;
+                    reject(error);
+                    throw error;
+                  })
+                );
+
+              this.subscription = obs.subscribe();
+            }
+          };
+        }
+      },
+      useResultCache: false,
+      maxSelect: '*',
+      matchListValue: true,
+      showListOnFocusIfEmpty: true
+    });
+    return acInstance;
+  }
+
+  /**
+   * Extracts autocomplete items from resource bundle
+   * @param bundle - resource bundle
+   * @param processedCodes - hash of processed items.
+   * Key is the EV name/description value; value is an array of EV IDs.
+   * Multiple EVs with the same name/description will be grouped.
+   * @param selectedCodes - already selected items
+   */
+  getAutocompleteItems_EV(
+    bundle: Bundle,
+    processedCodes: { [key: string]: string[] },
+    selectedCodes: Array<string>
+  ): ValueSetExpansionContains[] {
+    const result = [];
+    (bundle.entry || []).forEach((e) => {
+      const displayItem = e.resource[this.searchParameter];
+      if (processedCodes[displayItem]) {
+        processedCodes[displayItem].push(e.resource.id);
+      } else {
+        processedCodes[displayItem] = [e.resource.id];
+        if (selectedCodes.indexOf(displayItem) === -1) {
+          result.push({
+            display: displayItem,
+            code: processedCodes[displayItem]
+          });
+        }
+      }
+    });
+    return result;
+  }
+
+  /**
+   * Extracts autocomplete items from DbGap variable API response
+   * @param response - response from DbGap variable API
+   * The response takes the form of an array: first item is the total match on server;
+   * second item is an array of IDs, fourth item is an array of requested property (name/description).
+   * @param selectedCodes - already selected items
+   */
+  getAutocompleteItems_EV_dbgapVariableApi(
+    response: any,
+    selectedCodes: Array<string>
+  ): ValueSetExpansionContains[] {
+    if (!response[1]?.length) {
+      return [];
+    }
+    const result = [];
+    for (let i = 0; i < response[1].length; i++) {
+      const displayItem = response[3][i][0];
+      const id = AutocompleteParameterValueComponent.getEvIdFromDbgapVariableApi(
+        response[1][i]
+      );
+      const duplicateDisplayItem = result.find(
+        (x) => x.display === displayItem
+      );
+      if (duplicateDisplayItem) {
+        duplicateDisplayItem.code.push(id);
+      } else if (selectedCodes.indexOf(displayItem) === -1) {
+        result.push({
+          display: displayItem,
+          code: [id]
+        });
+      }
+    }
+    return result;
+  }
+
+  /**
    * Extracts autocomplete items from resource bundle
    * @param bundle - resource bundle
    * @param filterText - text in autocomplete field
@@ -407,6 +699,23 @@ export class AutocompleteParameterValueComponent
       );
       return acc;
     }, []);
+  }
+
+  /**
+   * Get 'q' params value for DbGap varaible API query.
+   * e.g. study_id:phs002410*, study_id:(phs002410*%20OR%20phs002409*)
+   * @private
+   */
+  private getDbgapEvResearchStudyParam(): string {
+    if (!this.researchStudy.myStudyIds.length) {
+      return '';
+    }
+    if (this.researchStudy.myStudyIds.length === 1) {
+      return `study_id:${this.researchStudy.myStudyIds[0]}*`;
+    }
+    return `study_id:(${this.researchStudy.myStudyIds
+      .map((id) => id + '*')
+      .join(' OR ')})`;
   }
 
   /**
@@ -463,5 +772,13 @@ export class AutocompleteParameterValueComponent
           return value.code ? [value] : value.coding || [];
         })
       );
+  }
+
+  getAriaLabel(): string {
+    return this.resourceType === this.EVIDENCEVARIABLE
+      ? `select Evidence Variables by ${this.searchParameter}`
+      : this.searchParameter === 'code'
+      ? `${this.resourceType} codes from FHIR server`
+      : 'Search parameter value';
   }
 }