+++ src/app/shared/query-params/query-params.service.ts
@@ -23,6 +23,9 @@ export class QueryParamsService {
    * @param value - search parameter value
    */
   getQueryParam(resourceType: string, value: SearchParameter): string {
+    if (resourceType === 'EvidenceVariable') {
+      return `&evidencevariable=${this.getEvidenceVariableIds(value)}`;
+    }
     const selectedParameter = this.definitions.resources[
       resourceType
     ]?.searchParameters.find((p) => p.element === value?.element);
@@ -73,6 +76,24 @@ export class QueryParamsService {
         : '';
     }
     return `&${selectedParameter.element}=${value.value}`;
+  }
+
+  /**
+   * Returns comma separated list of EV full URLs, to be used as query param
+   * for the EV search parameter.
+   * @param value search parameter value
+   * @private
+   */
+  private getEvidenceVariableIds(value: SearchParameter): string {
+    return value.value.codes
+      .map((codes: string[]) =>
+        codes
+          .map(
+            (c) => `${this.fhirBackend.serviceBaseUrl}/EvidenceVariable/${c}`
+          )
+          .join(',')
+      )
+      .join(',');
   }
 
   /**